+++ CHANGELOG.md
@@ -7,6 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Export PDF button that saves the PDF with selected layers and stitched pages
+- Save visible layers by file per browser/device
+- Save position by file per browser/device
+- Status icons that indicate when a PDF is loading and for when line thickness is updated
+
+### Changed
+
+- Always show stitch menu button regardless of the number of pages in the PDF
+- Disabled rendering of PDF links
+
+### Fixed
+
+- Close overlay options menu on click outside
+- Warning no longer flashes when dragging on Android, Chromebook, and in Firefox
+
 ## [1.0.2] - 2024-05-25
 
 ### Changed